update: enable sudo guards and update tests
