+++ frame/system/call_authorize_upgrade.go
@@ -74,12 +74,10 @@ func (_ callAuthorizeUpgrade) PaysFee(baseWeight primitives.Weight) primitives.P
 }
 
 func (c callAuthorizeUpgrade) Dispatch(origin primitives.RuntimeOrigin, args sc.VaryingData) (primitives.PostDispatchInfo, error) {
-	// TODO: enable once 'sudo' module is implemented
-	//
-	// err := EnsureRoot(origin)
-	// if err != nil {
-	// 	return err
-	// }
+	err := EnsureRoot(origin)
+	if err != nil {
+		return primitives.PostDispatchInfo{}, err
+	}
 
 	codeHash := args[0].(primitives.H256)
