+++ frame/system/call_set_code_without_checks.go
@@ -83,16 +83,14 @@ func (_ callSetCodeWithoutChecks) PaysFee(baseWeight primitives.Weight) primitiv
 }
 
 func (c callSetCodeWithoutChecks) Dispatch(origin primitives.RuntimeOrigin, args sc.VaryingData) (primitives.PostDispatchInfo, error) {
-	// TODO: enable once 'sudo' module is implemented
-	//
-	// err := EnsureRoot(origin)
-	// if err != nil {
-	// 	return primitives.PostDispatchInfo{}, err
-	// }
+	err := EnsureRoot(origin)
+	if err != nil {
+		return primitives.PostDispatchInfo{}, err
+	}
 
 	codeBlob := args[0].(sc.Sequence[sc.U8])
 
-	err := c.hookOnSetCode.SetCode(codeBlob)
+	err = c.hookOnSetCode.SetCode(codeBlob)
 	if err != nil {
 		return primitives.PostDispatchInfo{}, err
 	}
