+++ frame/system/call_set_heap_pages.go
@@ -83,12 +83,10 @@ func (_ callSetHeapPages) PaysFee(baseWeight primitives.Weight) primitives.Pays 
 }
 
 func (c callSetHeapPages) Dispatch(origin primitives.RuntimeOrigin, args sc.VaryingData) (primitives.PostDispatchInfo, error) {
-	// TODO: enable once 'sudo' module is implemented
-	//
-	// err := EnsureRoot(origin)
-	// if err != nil {
-	// 	return primitives.PostDispatchInfo{}, err
-	// }
+	err := EnsureRoot(origin)
+	if err != nil {
+		return primitives.PostDispatchInfo{}, err
+	}
 
 	pages := args[0].(sc.U64)
