+++ frame/system/call_set_storage.go
@@ -77,12 +77,10 @@ func (_ callSetStorage) PaysFee(baseWeight primitives.Weight) primitives.Pays {
 }
 
 func (c callSetStorage) Dispatch(origin primitives.RuntimeOrigin, args sc.VaryingData) (primitives.PostDispatchInfo, error) {
-	// TODO: enable once 'sudo' module is implemented
-	//
-	// err := EnsureRoot(origin)
-	// if err != nil {
-	// 	return primitives.PostDispatchInfo{}, err
-	// }
+	err := EnsureRoot(origin)
+	if err != nil {
+		return primitives.PostDispatchInfo{}, err
+	}
 
 	items := args[0].(sc.Sequence[KeyValue])
