+++ runtime/templates/poa/system_kill_storage_test.go
@@ -18,6 +18,10 @@ func Test_KillStorage_DispatchOutcome(t *testing.T) {
 	runtimeVersion, err := rt.Version()
 	assert.NoError(t, err)
 
+	// Set Sudo Key
+	err = (*storage).Put(append(testhelpers.KeySudoHash, testhelpers.KeyKeyHash...), signature.TestKeyringPairAlice.PublicKey)
+	assert.NoError(t, err)
+
 	testhelpers.InitializeBlock(t, rt, testhelpers.ParentHash, testhelpers.StateRoot, testhelpers.ExtrinsicsRoot, testhelpers.BlockNumber)
 
 	keys := [][]byte{
@@ -25,7 +29,10 @@ func Test_KillStorage_DispatchOutcome(t *testing.T) {
 		[]byte("testkey2"),
 	}
 
-	call, err := ctypes.NewCall(metadata, "System.kill_storage", keys)
+	callArg, err := ctypes.NewCall(metadata, "System.kill_storage", keys)
+	assert.NoError(t, err)
+
+	call, err := ctypes.NewCall(metadata, "Sudo.sudo", callArg)
 	assert.NoError(t, err)
 
 	extrinsic := ctypes.NewExtrinsic(call)
